Ui fixes updates
